chore: Make Docker image smaller & faster to build
First, we split build targets:
- for NPM, we want dual ESM/CJS + type definitions
- For Docker, we only need ESM

We can also leverage the PNPM in-memory cache with pnpm fetch,
to only run the deps install step when the lockfile changes
(not the individual package.json)

We can further thin the resulting image by pruning dev dependencies.
This involves moving the Jest config to devDeps, and moving
zx back to runtime deps, as we need it to run database ops.