fix: Use hex encoding to serialise ciphers
We compute a payload fingerprint by feeding it
a JSON-serialised cipher, which was using base64
to serialise Uint8Arrays. Since there are multiple base64 outputs
to represent the same input binary data, this could lead to
attacks where the same cipher produces different fingerprints.

Although the check involves recomputing the serialised cipher
and comparing the output, which would fail if one the sides had
been tampered with, this kind of non-determinism is not desired.

Using a hex encoding is deterministic and doesn't suffer from
this issue, at the cost of a few extra bytes of payload ciphertext.