ref(crypto): Derive identity from main key
Identity keys are now derived from the userId and main key.
This allows a login procedure without involving the server
(everything happens locally).

In addition, a signature is added as proof of ownership of
a given identity (signing userId and the sharing public key,
so that anyone can verify an identity).

Note: breaking changes in the data model require resetting
and reseeding the database.