ref(crypto): Use derived keys to encrypt keychain items
Rather than using the same key for all items' name & payload,
derive two keys (one for each field) per item, based on a
random subkey index stored in the database.

Since there are only 2^31 possible subkeys, collisions are possible
when a single user has a very large number of keys, though in practice
it does not weaken the encryption model as long as ChaCha20 is secure.