chore: Don't publish server code to NPM
Instead, run semantic-release on private packages,
which will not publish them on NPM, but allow
using the `publish` lifecycle hook on our custom plugin.

There, we can detect an intent to release the server,
which we store into the `server-needs-publishing` file
which Docker can pick up and decide whether to proceed.