fix: Prefix key "name" with randomness
Key injection becomes trivial if the name fingerprint can be reversed,
so the "name" needs to be sufficiently random to make this impractical.

The client API now exposes the passed parameter as a "label", also
accessible in the key metadata, and the actual name (which is
fingerprinted and encrypted) is prefixed with 32 bytes of randomness.

Devtools have been updated to show only the label in the key list,
and the whole "name" in the details.