feat: Authentication with OPAQUE, devices & TLS
Authentication is now done through an OPAQUE
key agreement to:
1. define an agreed session ID
2. obtain an export key to wrap the main key, per device

Each device has its set of credentials (user ID and device secret),
which are used for the OPAQUE key agreement.
Signup automatically enrolls the currently used device, and
enrollement of new devices can be done on an authenticated device.

The public key signature scheme is kept for authenticating requests,
though stateful sessions have been moved to Redis to both allow
expiration and listing for auditing.

Also, an end-to-end TLS connection is now enforced, with the help
of `mkcert` to establish localhost-bound certificates for development.
This ensures there can be no MitM on the transport side.