fix: Update form signatures to prevent canonicalisation attacks
BREAKING CHANGE: Signature pre-hash algorithm now includes
length prefixes for the number of elements to hash,
and the size of each element.
Also, the form public key has been removed from the list of
signed elements. Since it's already keying the pre-hash, it's
not necessary to include it twice in the signature calculation.